[project] Create Akane project, containing AkaneCore, AkaneBindings and AkaneCollections frameworks
This approach is Carthage compatible by default
