+++ Collections/Collection/Template/TemplateComponentView.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import Akane
 
 /**
  Create templates for `ComponentView`(s)
