+++ Projects/AkaneBindings/AkaneBindings.h
@@ -1,0 +1,19 @@
+//
+//  AkaneBindings.h
+//  AkaneBindings
+//
+//  Created by Simone Civetta on 23/11/16.
+//  Copyright © 2016 AkaneCore. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for AkaneBindings.
+FOUNDATION_EXPORT double AkaneBindingsVersionNumber;
+
+//! Project version string for AkaneBindings.
+FOUNDATION_EXPORT const unsigned char AkaneBindingsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AkaneBindings/PublicHeader.h>
+
+
