+++ Projects/AkaneCollections/AkaneCollections.h
@@ -1,0 +1,19 @@
+//
+//  AkaneCollections.h
+//  AkaneCollections
+//
+//  Created by Simone Civetta on 23/11/16.
+//  Copyright © 2016 AkaneCore. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for AkaneCollections.
+FOUNDATION_EXPORT double AkaneCollectionsVersionNumber;
+
+//! Project version string for AkaneCollections.
+FOUNDATION_EXPORT const unsigned char AkaneCollectionsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AkaneCollections/PublicHeader.h>
+
+
