--- Projects/All/AkaneAll/AkaneAll.h
@@ -1,19 +1,0 @@
-//
-//  AkaneAll.h
-//  AkaneAll
-//
-//  Created by Simone Civetta on 22/11/16.
-//  Copyright © 2016 Akane. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for AkaneAll.
-FOUNDATION_EXPORT double AkaneAllVersionNumber;
-
-//! Project version string for AkaneAll.
-FOUNDATION_EXPORT const unsigned char AkaneAllVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <AkaneAll/PublicHeader.h>
-
-
