--- Projects/All/Podfile
@@ -1,17 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-
-target 'AkaneAll' do
-  use_frameworks!
-
-  podspec :path => "../../Akane"
-  pod 'Akane', :path => "../../"
-  pod 'Akane/Bindings', :path => "../../"
-  pod 'Akane/Collections', :path => "../../"
-
-  target 'AkaneAllTests' do
-  	pod "Quick"
-  	pod "Nimble", '~> 5.x'
-  end
-
-end
