--- Projects/Core/AkaneCore/AkaneCore.h
@@ -1,19 +1,0 @@
-//
-//  AkaneCore.h
-//  AkaneCore
-//
-//  Created by Simone Civetta on 22/11/16.
-//  Copyright © 2016 Akane. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for AkaneCore.
-FOUNDATION_EXPORT double AkaneCoreVersionNumber;
-
-//! Project version string for AkaneCore.
-FOUNDATION_EXPORT const unsigned char AkaneCoreVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <AkaneCore/PublicHeader.h>
-
-
