--- Projects/Core/Podfile
@@ -1,15 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-
-target 'AkaneCore' do
-  use_frameworks!
-
-  podspec :path => "../../Akane"
-  pod 'Akane/Core', :path => "../../"
-
-  target 'AkaneCoreTests' do
-  	pod "Quick"
-  	pod "Nimble", '~> 5.x'
-  end
-
-end
